Config Routines added to CoPylot
Created _config_properties and _config_xml in CoPylot. They are located
in CoPylot due to their generic implementation - they should not need to
be re-implemented for every component the way Deploy/Configure/Run
will/may be.

Also created a backup of pylot_cfg as I'm going to use the new one as a
clean slate version (getting Val to work cleanly before moving on).