Steward + Config added
pylot_cfg recreated as config, using a class and instantiating as
pylot_cfg. Still a temp solution.

Steward added. Also implemented as a class. Kind of awkward since it's
mostly function calls, but it requires a reference to a Pylot Config for
directory structure stuff (which I don't want to re-do here, mostly for
style). Should think on this design though.
Logging functionality added from Pylot4 steward.

pylot updated to initialize an instance of Steward and Pylot_Config
which can be passed along to tests.